feat: error state and move things to the form
