+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/UploadPhotosForm.tsx
@@ -3,8 +3,9 @@ import { captureMessage } from "@sentry/react-native"
 import { useFormikContext } from "formik"
 import { GlobalStore } from "lib/store/GlobalStore"
 import { showPhotoActionSheet } from "lib/utils/requestPhotos"
-import { Button, Flex, Spacer, Text } from "palette"
+import { Button, Flex, Spacer, Spinner, Text } from "palette"
 import React, { useEffect } from "react"
+import { Image } from "react-native"
 import { uploadImageAndPassToGemini } from "../../../Submission/uploadPhotoToGemini"
 import { Photo, PhotosFormModel } from "./validation"
 
@@ -34,7 +35,6 @@ export const UploadPhotosForm: React.FC<{ setPhotoUploadError: (arg: boolean) =>
           })
 
           photo.loading = false
-          photo.error = true
           setFieldValue("photos", [...values.photos, photo])
         } catch (error) {
           photo.loading = false
@@ -57,21 +57,109 @@ export const UploadPhotosForm: React.FC<{ setPhotoUploadError: (arg: boolean) =>
     }
   }
 
+  const handlePhotoDelete = (photo: Photo) => {
+    console.log({ deleted_photo: photo })
+  }
+
   return (
-    <Flex style={{ borderColor: "lightgray", borderWidth: 1 }} mt={4} mb={2} p={2} pt={3} pb={3}>
-      <Text variant="lg" color="black100" marginBottom={1}>
-        Add Files Here
-      </Text>
-      <Text variant="md" color="black60" marginBottom={1}>
-        Files Supported: JPG, PNG
-      </Text>
-      <Text variant="md" color="black60" marginBottom={3}>
-        Total Maximum Size: 30MB
-      </Text>
-      <Button onPress={handleAddPhotoPress} variant="outline" size="large" block>
-        Add Photo
-      </Button>
-      <Spacer mt={1} />
-    </Flex>
+    <>
+      <Flex style={{ borderColor: "lightgray", borderWidth: 1 }} mt={4} mb={2} p={2} pt={3} pb={3}>
+        <Text variant="lg" color="black100" marginBottom={1}>
+          Add Files Here
+        </Text>
+        <Text variant="md" color="black60" marginBottom={1}>
+          Files Supported: JPG, PNG
+        </Text>
+        <Text variant="md" color="black60" marginBottom={3}>
+          Total Maximum Size: 30MB
+        </Text>
+        <Button onPress={handleAddPhotoPress} variant="outline" size="large" block>
+          Add Photo
+        </Button>
+        <Spacer mt={1} />
+      </Flex>
+
+      {values.photos.map((photo: Photo, idx: number) => {
+        if (photo?.loading) {
+          return <PhotoThumbnailLoadingState key={idx} />
+        } else if (photo?.error) {
+          return <PhotoThumbnailErrorState key={idx} />
+        }
+
+        return (
+          <PhotoThumbnailSuccessState
+            key={idx}
+            photo={photo}
+            handlePhotoDelete={handlePhotoDelete}
+          />
+        )
+      })}
+    </>
+  )
+}
+
+const PhotoThumbnailSuccessState: React.FC<{
+  photo: Photo
+  handlePhotoDelete: (arg: Photo) => void
+}> = ({ photo, handlePhotoDelete }) => {
+  return (
+    <>
+      <Flex
+        p={1}
+        flexDirection="row"
+        justifyContent="space-between"
+        alignItems="center"
+        style={{ borderColor: "lightgray", borderWidth: 1, borderRadius: 4, height: 68 }}
+      >
+        <Image
+          resizeMode="contain"
+          // TODO: get uril from photo
+          source={{
+            uri: "https://i.picsum.photos/id/14/200/300.jpg?hmac=FMdb1SH_oeEo4ibDe66-ORzb8p0VYJUS3xWfN3h2qDU",
+          }}
+          style={{ height: 58, width: 70 }}
+          // TODO
+          testID="image"
+        />
+        {/* TODO: actual size */}
+        <Text>0.32mb</Text>
+        <Button variant="text" size="small" onPress={() => handlePhotoDelete(photo)}>
+          <Text style={{ textDecorationLine: "underline" }}>Delete</Text>
+        </Button>
+      </Flex>
+      <Spacer mt={2} />
+    </>
+  )
+}
+
+const PhotoThumbnailLoadingState = () => {
+  return (
+    <>
+      <Flex
+        p={1}
+        flexDirection="row"
+        justifyContent="center"
+        alignItems="center"
+        style={{ borderColor: "lightgray", borderWidth: 1, borderRadius: 4, height: 68 }}
+      >
+        <Spinner color="black60" />
+      </Flex>
+      <Spacer mt={2} />
+    </>
+  )
+}
+
+const PhotoThumbnailErrorState = () => {
+  return (
+    <>
+      <Flex
+        p={1}
+        alignItems="center"
+        style={{ borderColor: "red", borderWidth: 1, borderRadius: 4, height: 68 }}
+      >
+        <Text color="black60">Problem</Text>
+      </Flex>
+      <Spacer mt={2} />
+    </>
   )
 }
